refactor(core): flatten api/, split freeze by phase, rename to proposal_framing
The core/api/ namespace pulled no weight after GroupHandle wraps the
freeze + inbound entry points. Move the three files up: freeze gets
split into freeze/{round,apply}.rs (round selection + apply phase
separated by the explicit phase comments already in the file),
inbound stays as inbound.rs, and lifecycle.rs becomes
proposal_framing.rs (its actual scope — wire framing helpers, not
group lifecycle, which already lives elsewhere).
