+++ src/core/freeze/mod.rs
@@ -1,0 +1,14 @@
+//! Freeze round candidate processing, selection, and commit application.
+//!
+//! Two submodules:
+//! - [`round`] — public surface plus per-round setup and priority selection.
+//! - [`apply`] — phase-3 loop: per-candidate apply, staging, validation,
+//!   and post-commit bookkeeping.
+
+mod apply;
+mod round;
+
+pub use round::{
+    FreezeFinalizeResult, FreezeOutcome, compute_commit_hash, finalize_freeze_round,
+    process_commit_candidate,
+};
